refactor: Ladder 도메인 로직 패키지 분리
