+++ src/main/java/io/suhan/ladder/model/ladder/Connection.java
@@ -1,4 +1,4 @@
-package io.suhan.ladder.model;
+package io.suhan.ladder.model.ladder;
 
 public class Connection {
     private final int left;
